+++ src/api/plugins/video/warren_video/indicators.py
@@ -5,12 +5,13 @@ from typing import List
 from ralph.backends.http.lrs import BaseHTTP, LRSQuery
 from ralph.models.xapi.concepts.constants.video import RESULT_EXTENSION_TIME
 from ralph.models.xapi.concepts.verbs.scorm_profile import CompletedVerb
+from ralph.models.xapi.concepts.verbs.tincan_vocabulary import DownloadedVerb
 from ralph.models.xapi.concepts.verbs.video import PlayedVerb
 from warren.base_indicator import BaseIndicator, pre_process_statements
 from warren.filters import DatetimeRange
 from warren.models import XAPI_STATEMENT
 from warren_video.conf import settings as video_plugin_settings
-from warren_video.models import VideoViews
+from warren_video.models import VideoDownloads, VideoViews
 
 
 class DailyVideoViews(BaseIndicator):
@@ -185,6 +186,88 @@ class DailyCompletedVideoViews(BaseIndicator):
 
         # Calculate the total number of events
         indicator.total_views = len(flattened.index)
+        indicator.count_by_date = count_by_date.to_dict("records")
+
+        return indicator
+
+
+class DailyVideoDownloads(BaseIndicator):
+    """Daily Video Downloads indicator.
+
+    Defines the daily video downloads indicator. It is, for a given video, and
+    a date range, the total number of downloads, and the number of downloads per day.
+    """
+
+    def __init__(
+        self,
+        client: BaseHTTP,
+        video_uuid: str,
+        date_range: DatetimeRange,
+        is_unique: bool,
+    ):
+        """Instantiate the indicator with its parameters.
+
+        Args:
+            client: The LRS backend to query
+            video_uuid: The UUID of the video on which to compute the metric
+            date_range: The date range on which to compute the indicator. It has
+                2 fields, `since` and `until` which are dates or timestamps that must be
+                in ISO format (YYYY-MM-DD, YYYY-MM-DDThh:mm:ss.sss±hh:mm or
+                YYYY-MM-DDThh:mm:ss.sssZ")
+            is_unique: If true, multiple downloads by the same actor are counted as one
+        """
+        self.client = client
+        self.video_uuid = video_uuid
+        self.date_range = date_range
+        self.is_unique = is_unique
+
+    def get_lrs_query(self) -> LRSQuery:
+        """Returns the LRS query for fetching required statements."""
+        return LRSQuery(
+            query={
+                "verb": DownloadedVerb().id,
+                "activity": self.video_uuid,
+                "since": self.date_range.since.isoformat(),
+                "until": self.date_range.until.isoformat(),
+            }
+        )
+
+    def fetch_statements(self) -> List[XAPI_STATEMENT]:
+        """Executes the LRS query to obtain statements required for this indicator."""
+        return list(
+            self.client.read(
+                target=self.client.statements_endpoint, query=self.get_lrs_query()
+            )
+        )
+
+    def compute(self) -> VideoDownloads:
+        """Fetches statements and computes the current indicator.
+
+        Fetches the statements from the LRS, filters and aggregates them to return the
+        number of video downloads per day.
+        """
+        indicator = VideoDownloads()
+        raw_statements = self.fetch_statements()
+        if not raw_statements:
+            return indicator
+        preprocessed_statements = pre_process_statements(raw_statements)
+
+        # If the `is_unique` filter is selected, filter out rows with duplicate
+        # `actor.account.name`
+        if self.is_unique:
+            preprocessed_statements.drop_duplicates(subset="actor.uuid", inplace=True)
+
+        # Group by day and calculate sum of downloads per day
+        count_by_date = (
+            preprocessed_statements.groupby(preprocessed_statements["date"])
+            .count()
+            .reset_index()
+            .rename(columns={"id": "count"})
+            .loc[:, ["date", "count"]]
+        )
+
+        # Calculate the total number of downloads
+        indicator.total_downloads = len(preprocessed_statements.index)
         indicator.count_by_date = count_by_date.to_dict("records")
 
         return indicator